+++ Changelog.md
@@ -1,8 +1,8 @@
 # FactionsPlus 0.7.1
 * Added configuration option: noDisbandIfPeaceful
+* Added configuration option: peacefulChestProtect
 
 # FactionsPlus 0.7.0
-
 * Huge code clean up
 * Fixed a small issue with not ensuring FactionsPlusUpdate was not running
 * Now using MassiveCore